+++ spec/rubocop/cop/rspec/described_class_spec.rb
@@ -267,6 +267,18 @@ RSpec.describe RuboCop::Cop::RSpec::DescribedClass, :config do
         end
       RUBY
     end
+
+    it 'ignores if `described_class` is a part of the constant' do
+      expect_no_offenses(<<-RUBY)
+        module SomeGem
+          describe VERSION do
+            it 'returns proper version string' do
+              expect(described_class::STRING).to eq('1.1.1')
+            end
+          end
+        end
+      RUBY
+    end
   end
 
   context 'when EnforcedStyle is :explicit' do